requirements.txt
bitarray==0.8.3
imageio==2.4.1
numpy==1.15.4
Pillow==5.3.0
